+++ CHANGES.rst
@@ -10,6 +10,10 @@ CHANGES
 
 - Declare Python 3.9 compatibility
 
+- Small improvement in `pjpersist.zope.container.PJContainer._load_one`
+  Get the local cache just once, since the `_cache` property became
+  a bit more expensive.
+
 
 3.0.0 (2021-02-22)
 ------------------